templates/backend-python/pyproject.toml
[project]
name = "__PROJECT_NAME__"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "sqlalchemy[asyncio]==2.0.36",
    "asyncpg==0.29.0",
    "pydantic-settings==2.5.2",
]

[project.optional-dependencies]
dev = [
    "ruff==0.6.9",
    "pytest==8.3.3",
    "httpx==0.27.2",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
